make it possible to count runs/pairs and flushes correctly
Counting runs and pairs means the hand needs to be sorted.
Counting flushes means we need to distinguish the starter card --
which formerly relied on the starter being last, but that breaks sort
order!

Instead, add an explicit field to hand_t to record which card is the
starter (if any).